4. Refactor Album List, after homeWork4 review.
