+++ src/entities/album/ui/Album.module.scss
@@ -1,0 +1,26 @@
+.albumsItem {
+  border-radius: 8px;
+  inline-size: fit-content;
+  background-color: var(--comment-background-color);
+  padding:10px;
+  cursor:pointer;
+}
+
+.albumCover {
+  border-radius: 8px;
+  inline-size: 22vw;
+  aspect-ratio: 1/1;
+  object-fit: cover;
+
+  @media (max-width:1080px) {
+    inline-size: 33vw;
+  }
+
+  @media (max-width:750px) {
+    inline-size: 66vw;
+  }
+}
+
+.albumsItemTitle {
+  color: var(--text-color);
+}
