+++ src/pages/user_albums/user_albums.module.scss
@@ -14,30 +14,3 @@
   border-radius: 8px;
   padding: 0.5rem;
 }
-
-.albumsItem {
-  border-radius: 8px;
-  inline-size: fit-content;
-  background-color: var(--comment-background-color);
-  padding:10px;
-  cursor:pointer;
-}
-
-.albumCover {
-  border-radius: 8px;
-  inline-size: 22vw;
-  aspect-ratio: 1/1;
-  object-fit: cover;
-
-  @media (max-width:1080px) {
-    inline-size: 33vw;
-  }
-
-  @media (max-width:750px) {
-    inline-size: 66vw;
-  }
-}
-
-.albumsItemTitle {
-  color: var(--text-color);
-}
